Fixes incorrect root joint detection bug in glTF importer.
glTF 2.0 specification requires that all joints referenced by a skin
have a common root, but does not require that this root is referenced
by the skin directly. glTF importer assumed that, which caused invalid
joint to be chosen as the root of the skeleton.

To fix this entire joint forest for the scene must be reversed, so
FindSkinRootJointIndices method is also slightly refactored to this
only once.